Added a win lose logic for endscreen
Implemented logic so that when the end screen is activated it checks to see if the player is at the end position if so the winning screen is displayed if not the losing screen is displayed.  Also had to implement variable to the gameboard class so that way the end position is known for all boards.